// src/types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  public: {
    Tables: {
// ...
      complaint_messages: {
        Row: {
          complaint_id: string
          created_at: string
          id: string
          message: string
          sender_id: string
          sender_role: string
        }
        Insert: {
          complaint_id: string
          created_at?: string
          id?: string
          message: string
          sender_id: string
          sender_role: string
        }
        Update: {
          complaint_id?: string
          created_at?: string
          id?: string
          message?: string
          sender_id?: string
          sender_role?: string
        }
        Relationships: [
          {
            foreignKeyName: "complaint_messages_complaint_id_fkey"
            columns: ["complaint_id"]
            isOneToOne: false
            referencedRelation: "complaints"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      order_items: {
        Row: {
          id: string
          image_url: string | null
          order_id: string
          price: number
          product_id: string | null
          product_name: string
          quantity: number
          sku: string
          subtotal: number
          variant_id: string | null
          variant_name: string
          weight: number
        }
        Insert: {
          id?: string
          image_url?: string | null
          order_id: string
          price: number
          product_id?: string | null
          product_name: string
          quantity: number
          sku: string
          subtotal: number
          variant_id?: string | null
          variant_name: string
          weight: number
        }
        Update: {
          id?: string
          image_url?: string | null
          order_id?: string
          price?: number
          product_id?: string | null
          product_name?: string
          quantity?: number
          sku?: string
          subtotal?: number
          variant_id?: string | null
          variant_name?: string
          weight?: number
        }
        Relationships: [
          {
            foreignKeyName: "order_items_order_id_fkey"
            columns: ["order_id"]
            isOneToOne: false
            referencedRelation: "orders"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "order_items_variant_id_fkey"
            columns: ["variant_id"]
            isOneToOne: false
            referencedRelation: "product_variants"
            referencedColumns: ["id"]
          },
        ]
      }
      order_status_history: {
        Row: {
          changed_by: string | null
          created_at: string
          id: string
          note: string | null
          order_id: string
          status: Database["public"]["Enums"]["order_status"]
        }
        Insert: {
          changed_by?: string | null
          created_at?: string
          id?: string
          note?: string | null
          order_id: string
          status: Database["public"]["Enums"]["order_status"]
        }
        Update: {
          changed_by?: string | null
          created_at?: string
          id?: string
          note?: string | null
          order_id?: string
          status?: Database["public"]["Enums"]["order_status"]
        }
        Relationships: [
          {
            foreignKeyName: "order_status_history_changed_by_fkey"
            columns: ["changed_by"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "order_status_history_order_id_fkey"
            columns: ["order_id"]
            isOneToOne: false
            referencedRelation: "orders"
            referencedColumns: ["id"]
          },
        ]
      }
      orders: {
        Row: {
          admin_viewed_at: string | null
          app_fee: number
          coupon_code: string | null
          coupon_id: string | null
          courier_company: string | null
          courier_etd: string | null
          courier_service: string | null
          created_at: string
          delivered_at: string | null
          discount_amount: number
          id: string
          notes: string | null
          order_number: string
          recipient_name: string
          recipient_phone: string
          refund_account_name: string | null
          refund_account_number: string | null
          refund_bank_name: string | null
          shipping_address: string
          shipping_city: string
          shipping_cost: number
          shipping_district: string
          shipping_insurance: number
          shipping_lat: number | null
          shipping_lng: number | null
          shipping_postal: string
          shipping_province: string
          status: Database["public"]["Enums"]["order_status"]
          subtotal: number
          total: number
          updated_at: string
          user_id: string | null
        }
        Insert: {
          admin_viewed_at?: string | null
          app_fee?: number
          coupon_code?: string | null
          coupon_id?: string | null
          courier_company?: string | null
          courier_etd?: string | null
          courier_service?: string | null
          created_at?: string
          delivered_at?: string | null
          discount_amount?: number
          id?: string
          notes?: string | null
          order_number?: string
          recipient_name: string
          recipient_phone: string
          refund_account_name?: string | null
          refund_account_number?: string | null
          refund_bank_name?: string | null
          shipping_address: string
          shipping_city: string
          shipping_cost?: number
          shipping_district: string
          shipping_insurance?: number
          shipping_lat?: number | null
          shipping_lng?: number | null
          shipping_postal: string
          shipping_province: string
          status?: Database["public"]["Enums"]["order_status"]
          subtotal: number
          total: number
          updated_at?: string
          user_id?: string | null
        }
        Update: {
          admin_viewed_at?: string | null
          app_fee?: number
          coupon_code?: string | null
          coupon_id?: string | null
          courier_company?: string | null
          courier_etd?: string | null
          courier_service?: string | null
          created_at?: string
          delivered_at?: string | null
          discount_amount?: number
          id?: string
          notes?: string | null
          order_number?: string
          recipient_name?: string
          recipient_phone?: string
          refund_account_name?: string | null
          refund_account_number?: string | null
          refund_bank_name?: string | null
          shipping_address?: string
          shipping_city?: string
          shipping_cost?: number
          shipping_district?: string
          shipping_insurance?: number
          shipping_lat?: number | null
          shipping_lng?: number | null
          shipping_postal?: string
          shipping_province?: string
          status?: Database["public"]["Enums"]["order_status"]
          subtotal?: number
          total?: number
          updated_at?: string
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "orders_coupon_id_fkey"
            columns: ["coupon_id"]
            isOneToOne: false
            referencedRelation: "coupons"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "orders_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      product_variants: {
        Row: {
          created_at: string
          height: number
          id: string
          image_id: string | null
          image_url: string | null
          is_active: boolean
          length: number
          name: string
          price: number
          product_id: string
          reserved: number
          sku: string
          stock: number
          updated_at: string
          weight: number
          width: number
        }
        Insert: {
          created_at?: string
          height?: number
          id?: string
          image_id?: string | null
          image_url?: string | null
          is_active?: boolean
          length?: number
          name: string
          price: number
          product_id: string
          reserved?: number
          sku: string
          stock?: number
          updated_at?: string
          weight: number
          width?: number
        }
        Update: {
          created_at?: string
          height?: number
          id?: string
          image_id?: string | null
          image_url?: string | null
          is_active?: boolean
          length?: number
          name?: string
          price?: number
          product_id?: string
          reserved?: number
          sku?: string
          stock?: number
          updated_at?: string
          weight?: number
          width?: number
        }
        Relationships: [
          {
            foreignKeyName: "product_variants_image_id_fkey"
            columns: ["image_id"]
            isOneToOne: false
            referencedRelation: "product_images"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "product_variants_product_id_fkey"
            columns: ["product_id"]
            isOneToOne: false
            referencedRelation: "products"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      stock_history: {
        Row: {
          changed_by: string | null
          created_at: string
          id: string
          note: string | null
          order_id: string | null
          quantity: number
          type: string
          variant_id: string
        }
        Insert: {
          changed_by?: string | null
          created_at?: string
          id?: string
          note?: string | null
          order_id?: string | null
          quantity: number
          type: string
          variant_id: string
        }
        Update: {
          changed_by?: string | null
          created_at?: string
          id?: string
          note?: string | null
          order_id?: string | null
          quantity?: number
          type?: string
          variant_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "stock_history_changed_by_fkey"
            columns: ["changed_by"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "stock_history_order_id_fkey"
            columns: ["order_id"]
            isOneToOne: false
            referencedRelation: "orders"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "stock_history_variant_id_fkey"
            columns: ["variant_id"]
            isOneToOne: false
            referencedRelation: "product_variants"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      is_admin: { Args: never; Returns: boolean }
    }
    Enums: {
      order_status:
        | "pending_payment"
        | "paid"
        | "processing"
        | "shipped"
        | "delivered"
        | "completed"
        | "cancelled"
        | "refunded"
      payment_status:
        | "pending"
        | "paid"
        | "failed"
        | "expired"
        | "cancelled"
        | "refunded"
        | "challenge"
      shipment_status:
        | "pending"
        | "confirmed"
        | "allocated"
        | "picking_up"
        | "picked"
        | "dropping_off"
        | "delivered"
        | "rejected"
        | "cancelled"
        | "returned"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}